```
BenchmarkTrebuchetPart1-10    	   68078	     16971 ns/op
BenchmarkTrebuchetPart2-10    	   33944	     30591 ns/op
```
----
Starting point: https://github.com/pivovarit/AoC_2023_go/commit/a6b330dd21625155f0190c06ada08eec7ea535b8

```
BenchmarkTrebuchetPart1-10    	   34257	     34038 ns/op
BenchmarkTrebuchetPart2-10    	     135	   8590369 ns/op
```
Iteration 1: https://github.com/pivovarit/AoC_2023_go/commit/c1b90513d7e62aeb46331e7c1903324df32c9b9d 
```
BenchmarkTrebuchetPart1-10    	   34888	     34297 ns/op
BenchmarkTrebuchetPart2-10    	    1473	    794652 ns/op
```
Iteration 2: https://github.com/pivovarit/AoC_2023_go/commit/6f4a14bc34d3f942a4708d116de729b02c29755f
```
BenchmarkTrebuchetPart1-10    	   45985	     23562 ns/op
BenchmarkTrebuchetPart2-10    	    5109	    228439 ns/op
```
Iteration 3: https://github.com/pivovarit/AoC_2023_go/commit/a3ee47596331f54bc45c4e651d761ddca740ceeb
```
BenchmarkTrebuchetPart1-10    	   46712	     23556 ns/op
BenchmarkTrebuchetPart2-10    	   22686	     50887 ns/op
```
Iteration 4: https://github.com/pivovarit/AoC_2023_go/commit/c8cd511e0e47a159030448a2b915850fab751b05
```
BenchmarkTrebuchetPart1-10    	   60452	     16862 ns/op
BenchmarkTrebuchetPart2-10    	   24405	     46625 ns/op
```
Iteration 5: https://github.com/pivovarit/AoC_2023_go/commit/5ddf9f14a3b215770c4fd63cc6ab5f828835f514
```
BenchmarkTrebuchetPart1-10    	   61749	     16854 ns/op
BenchmarkTrebuchetPart2-10    	   25545	     44756 ns/op
```